Fixing trac #28989
Definition of cycle(self) when proper=True:
Currently, if the cycle has odd length, then the proper cycle is set to be equal to the cycle. 
However, according to Prop 6.10.5 in Buchmann Vollmer, if the cycle has odd length, then the proper cycle consists of the cycle and its tau-images.
Instead of computing the proper cycle from the (improper) cycle, I suggest to iterate _Rho() if proper=True and _RhoTau() if proper=False.